[ANDROSDK-1998] move api logic to network layer
