+++ core/src/test/java/org/hisp/dhis/android/core/arch/api/fields/internal/FieldsStringGenerationShould.kt
@@ -41,7 +41,6 @@ import org.hisp.dhis.android.core.event.internal.EventQueryCriteriaFields
 import org.hisp.dhis.android.core.event.internal.NewEventFields
 import org.hisp.dhis.android.core.fileresource.internal.FileResourceFields
 import org.hisp.dhis.android.core.legendset.internal.LegendFields
-import org.hisp.dhis.android.core.map.layer.internal.externalmap.ExternalMapLayerFields
 import org.hisp.dhis.android.core.note.internal.NewNoteFields
 import org.hisp.dhis.android.core.note.internal.NoteFields
 import org.hisp.dhis.android.core.organisationunit.internal.OrganisationUnitFields
@@ -105,6 +104,7 @@ import org.hisp.dhis.android.network.dataset.SectionFields
 import org.hisp.dhis.android.network.datavalue.DataValueFields
 import org.hisp.dhis.android.network.eventfilter.EventFilterFields
 import org.hisp.dhis.android.network.expressiondimensionitem.ExpressionDimensionItemFields
+import org.hisp.dhis.android.network.externalmaplayer.ExternalMapLayerFields
 import org.hisp.dhis.android.network.indicator.IndicatorFields
 import org.hisp.dhis.android.network.indicatortype.IndicatorTypeFields
 import org.hisp.dhis.android.network.legendset.LegendSetFields
